Setup boilerplate script
Closes #7 and #4
