+++ androidApp/src/main/java/com/mirego/kmp/boilerplate/Greeting.kt
@@ -1,26 +1,12 @@
-package com.mirego.kmp.boilerplate.android
+package com.mirego.kmp.boilerplate
 
-import android.os.Bundle
-import androidx.activity.compose.setContent
-import androidx.appcompat.app.AppCompatActivity
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.tooling.preview.Preview
-import com.mirego.kmp.boilerplate.Greeting
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flowOf
-
-
-class MainActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContent {
-            Greeting(textFlow = Greeting().greeting())
-        }
-    }
-}
 
 @Composable
 fun Greeting(textFlow: Flow<String>) {
