+++ androidApp/src/main/java/com/mirego/kmp/boilerplate/MainActivity.kt
@@ -1,0 +1,15 @@
+package com.mirego.kmp.boilerplate
+
+import android.os.Bundle
+import androidx.activity.compose.setContent
+import androidx.appcompat.app.AppCompatActivity
+
+
+class MainActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContent {
+            Greeting(textFlow = Greeting().greeting())
+        }
+    }
+}
