+++ ios/Podfile
@@ -1,6 +1,5 @@
 target 'iosApp' do
   use_frameworks!
   platform :ios, '14.1'
-  pod 'shared', :path => '../shared'
-  pod 'SwiftLint'
+  pod 'Shared', :path => '../shared'
 end
