+++ ios/iosApp/Assets.xcassets/AppIcon.appiconset/Contents.json
